More fixes when deleteing entities.

+++ src/Plugin/QueueWorker/DeletedWorkspaceQueue.php
@@ -5,6 +5,7 @@ namespace Drupal\multiversion\Plugin\QueueWorker;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Core\Queue\QueueWorkerBase;
+use Drupal\Core\Utility\Error;
 use Drupal\multiversion\Entity\Storage\ContentEntityStorageInterface;
 use Drupal\multiversion\Entity\Workspace;
 use Drupal\multiversion\Workspace\WorkspaceManagerInterface;
@@ -27,11 +28,17 @@ class DeletedWorkspaceQueue extends QueueWorkerBase implements ContainerFactoryP
   protected $entityTypeManager;
 
   /**
+   * @var \Drupal\multiversion\Workspace\WorkspaceManagerInterface
+   */
+  private $workspaceManager;
+
+  /**
    * {@inheritdoc}
    */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, EntityTypeManagerInterface $entity_type_manager) {
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, EntityTypeManagerInterface $entity_type_manager, WorkspaceManagerInterface $workspace_manager) {
     parent::__construct($configuration, $plugin_id, $plugin_definition);
     $this->entityTypeManager = $entity_type_manager;
+    $this->workspaceManager = $workspace_manager;
   }
 
   /**
@@ -53,14 +60,30 @@ class DeletedWorkspaceQueue extends QueueWorkerBase implements ContainerFactoryP
    * @throws \Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    * @throws \Drupal\Core\Entity\EntityStorageException
+   * @throws \Exception
    */
   public function processItem($data) {
     $storage = $this->entityTypeManager->getStorage($data['entity_type_id']);
-    if ($storage instanceof ContentEntityStorageInterface) {
+    if ($storage instanceof ContentEntityStorageInterface && !empty($data['workspace_id'])) {
+      $workspace = Workspace::load($data['workspace_id']);
+      if ($workspace) {
+        $this->workspaceManager->setActiveWorkspace($workspace);
+      }
       $original_storage = $storage->getOriginalStorage();
       $entity = $original_storage->load($data['entity_id']);
       if ($entity) {
-        $original_storage->delete([$entity]);
+        try {
+          $original_storage->delete([$entity]);
+        }
+        catch (\Throwable $e) {
+          $arguments = Error::decodeException($e) + ['%uuid' => $entity->uuid()];
+          $message = t('%type: @message in %function (line %line of %file). The error occurred while deleting the entity with the UUID: %uuid.', $arguments);
+          throw new \Exception($message);
+        }
+      }
+      $default_workspace = Workspace::load(\Drupal::getContainer()->getParameter('workspace.default'));
+      if ($default_workspace) {
+        $this->workspaceManager->setActiveWorkspace($default_workspace);
       }
     }
     elseif ($data['entity_type_id'] == 'workspace') {
